+++ currencies/src/mock.rs
@@ -38,7 +38,7 @@ parameter_types! {
 	pub const AvailableBlockRatio: Perbill = Perbill::one();
 }
 
-pub type AccountId = u64;
+pub type AccountId = u128;
 impl frame_system::Trait for Runtime {
 	type Origin = Origin;
 	type Call = ();
